Experiment with instanceof folding
MultiResultSCRules was split into the interface and
implementation and a new experimental implementation of the rules was added.
This implementation branches when it is possible to fold with a substitution
that is not a renaming.
Also added the instanceof folding method (fold when the lower configuration
is an instance of the upper). The experimental implementation of the rules
should be used with this folding method.