New method of graph building and processing
It is temporarily called Messy. The idea is to build a graph
without repeatition of configurations, i.e. each configuration
corresponds to a single node. This enables us to perform some
interesting algorithms like residuation with caching (function residuate2).
All this stuff is written in the imperative style, sorry.

Also added a he-based whistle for the higher language and fixed something.